refactor: useFunnel에서 useCallback 훅 사용으로 최적화
